Add secret value to service environment
Using the admin RDS secret for the sake of producing a solution. This
should never be done and a new secret only for the application should be
used. Unfortunately this can easily done in CDK and CloudFront so a more
complex solution should be used. This is explained in the solution.md
file